issue #116 and issue #70
