+++ CONTRIBUTING.md
@@ -1,0 +1,40 @@
+# Our Contribution Guidelines:
+
+## As a Team
+
+Be aware of Grace Shopper Related Objectives:
+
+- Project management and effective use of Git Pull Requests/Github.
+- Emphasize best practices
+- Practice using proper Git workflow (feature branching, pull requests, etc.)
+- Practice Agile team management
+  [gist](https://gist.github.com/khumphrey/855bbc17481a6b0866440df141a33407)
+- The group is responsible for coming up with descriptions for new tasks that need to be done when we have our standup and TaskMaster will create the new tasks.
+- members are responsible for looking at requirements and coming up with ideas.
+- members should be aware that Git Master and Task Master are coding & keeping up with their extra responsibilities.
+
+## Expectations
+
+### Group Members:
+
+- git-hygiene aware.
+- Update their task with comments and status:
+  - Being clear and concise and understandable.
+  - It should be easy for other group members to easily understand what you're working on and its status by looking at the project board.
+- link issue to pull request when initiating "Compare and Pull" in github.
+- Refer to github docs [here](./README.md) And first try to figure out how to properly complete your task.
+  -Then Ask gitMaster for help on this after that if you’re still unsure.
+  **Goal should be to keep Project board up to date so that we aren't communicating about project issues through discord and slack.**
+  _Suggestion_: Using Git discussions on Grace Shopper Repo
+
+### GitMaster:
+
+- is in charge of approving PR and checking that PR is linked to issue.
+- Can reject PR if it is not properly linked to the issue and/or is in some way incomplete/unclear gitMaster can make comment “missing issue reference” etc and send it back to the member.
+- GitMaster should not have to do more than ensuring the git workflow is clean and well kept.
+- Someone else will approve git Master’s pull requests.
+
+### TaskMaster
+
+- Responsible for assigning new tasks and double checking that group members are updating their status.
+- Task Master should also not have to do more than making sure people have tasks.
