+++ NORMS.md
@@ -1,0 +1,51 @@
+## Group Norms:
+
+[example](https://github.com/FullstackAcademy/2206-GHP-RM-WEB-FT/blob/main/03-senior-phase/team-contract-example.pdf)
+
+#### Roles
+
+- GitMaster
+- TaskMaster
+
+#### Daily Process:
+
+10 AM
+Check in around 4PM
+
+### Disagreements and Frustration:
+
+- Group comfort with using a new technology is important.
+- Pros and cons list, everyone involved, last case ask alec for input if its a technology choice.
+- Everyone’s input is important, but Majority rules.
+- Take a break
+- Be supportive and understanding.
+- Be respectful, if someone has an issue with another team member talk with that person about it respectfully.
+- Ask for help sooner rather than later
+- Communicate openly
+
+### Merges
+
+- Taken care of by GitMaster
+- Adhering to Contribution Guidelines will keep the merge conflicts away!
+
+### Asking for Help:
+
+- if a group member is struggling with a task they should tell the group.
+- If we’ve already asked for help within the group, and everyone is stumped, it’s ok to ask for instructor assistance!
+
+### Pairing:
+
+- Pairs decide how they want to accomplish tasks.
+- Switch after completing a requirement, component, etc
+- Partners should switch off 45-90 min, breaking in between.
+
+### Work After Hours
+
+- It’s ok to keep working on something minor in your free time
+- Morning merge meeting as part of the stand up
+- Respect different workflows.
+- If taking on a new task from our "todos", Communicate with the group and the taskMaster, so that they can update the project board and the group knows what you're working on.
+- Be aware of what files are being worked on by others.
+- Don't expect the gitmaster to approve pull requests after normal hours.
+- Talk about what we’ve been doing to stay on the same page.
+- Try and find discrete components or parts that you can work on if you do want to work asynchronously to minimize conflicts
