+++ README.md
@@ -1,29 +1,63 @@
-### Grace-Shopper
+## Grace-Shopper
 
-## project board best practices
+### For General Information:
 
-# For Collaborators:
+[Norms](./NORMS.md)
 
-- when you're assigned a task, it should be in the _"Todo"_ status group.
-- update your status from todo to _"in progress"_ when you begin working on a task.
-  \_Be cognizant of updating your status, and add any relevant comments to your task. i.e. something's not working - change the status to "bug" and be concise and clear about the problem, where it occurs, and suggested solution if any.
+### For Collaborators:
 
-# git checkout -b [newBranchName]
+[Contribution_Guidelines](./CONTRIBUTING.md)
 
-- _BranchName_ :
-  should be descriptive and short and related to the title of the issue you're working on.
+#### github links to documentation:
 
-_commit messages_:
-for your final commit before placing a pull request - it should have the same title as your issue (task) or you can manually link it, see the link below for linking a PR to an issue.
-\*\* you can change the title of the issue to make it shorter and then add in the description what it was.
-
-## github links to documentation:
+[git-guide](https://rogerdudler.github.io/git-guide/)
 
 _git commands for pushes and commits to remotes_
+
 https://docs.github.com/en/get-started/using-git/pushing-commits-to-a-remote-repository#pushing-tags
 
 _Linking a pull request to an issue_
 
-> https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword > _Task Master_ & _gitMaster_ > https://docs.github.com/en/issues/planning-and-tracking-with-projects/learning-about-projects/about-projects
+- https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
 
-> https://docs.github.com/en/issues/planning-and-tracking-with-projects/automating-your-project/using-the-built-in-automations
+[Projects](https://docs.github.com/en/issues/planning-and-tracking-with-projects/learning-about-projects/about-projects)
+
+[ProjectAutomation](https://docs.github.com/en/issues/planning-and-tracking-with-projects/automating-your-project/using-the-built-in-automations)
+
+### FSA suggested Group Work:
+
+#### Group Work
+
+In this project, you will work in teams of 4. In a team of four people A B C D, there are 3 unique pairing combinations: AB CD, AC BD, AD BC. As you work through the project, please continue to switch pairs on a feature by feature basis. If a feature is taking more than two days, please consult your assigned fellow or an instructor.
+
+Use Git to manage your project, and GitHub for easy collaboration. As needed, meet as a team to decide on the next reachable goals for your project and split them into achievable tasks. For a given task, we recommend the following workflow.
+
+[Suggested](https://learn.fullstackacademy.com/workshop/54b3de126adfdb0b0027d52f/content/54b3de6f6adfdb0b0027d559/text)
+
+For more on git workflow in Grace Shopper check out this
+[gist](https://gist.github.com/khumphrey/e9ccb8139312b8db9718cd91dc2d1357)
+
+#### Grace Shopper Milestones:
+
+The Grace Shopper workshop has a lot of stories, and it’s easy to get overwhelmed. These are the major features we’re looking for at each code review (🌟means it’s extra credit)
+
+**Catalog (due Code Review 1)**
+
+Customers can browse & filter by category
+Admins can add & edit products
+🌟Customers can do full text search on the catalog
+🌟Catalog is paginated
+
+**Cart (due Code Review 2)**
+
+Customers have a cart they can put stuff in while they shop
+Customers can see their order history
+🌟Customers always have their cart, on any device
+
+**Checkout (due during Presentation)**
+
+Customers can checkout, using a real payment api (probably Stripe)
+Admins can see pending orders that need to be shipped
+🌟Customers get email confirmations
+
+[Tiers](https://github.com/FullstackAcademy/2206-GHP-RM-WEB-FT/blob/main/03-senior-phase/Grace_Shopper_Tiers.md)
